backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
# psycopg2-binary
django-cors-headers
drf-spectacular
Pillow
python-dotenv
gunicorn
dj-database-url
requests